Navigation working again
