+++ DungeonsandDragons/app/src/main/java/com/example/dungeonsanddragons/LoginViewModel.kt
@@ -15,5 +15,4 @@ class LoginViewModel : ViewModel() {
     }
 
     val firebaseUserData = FirebaseUserLiveData()
-    }
 }
