+++ DungeonsandDragons/app/src/main/java/com/example/dungeonsanddragons/MainFragment.kt
@@ -1,7 +1,6 @@
 package com.example.dungeonsanddragons
 
 import android.app.Activity
-import android.content.Intent
 import android.os.Bundle
 import android.util.Log
 import androidx.fragment.app.Fragment
@@ -10,15 +9,16 @@ import android.view.View
 import android.view.ViewGroup
 import androidx.fragment.app.viewModels
 import androidx.lifecycle.Observer
+import androidx.navigation.NavController
 import androidx.navigation.fragment.findNavController
-import com.example.dungeonsanddragons.databinding.FragmentLoginBinding
+import com.example.dungeonsanddragons.databinding.FragmentMainBinding
 import com.firebase.ui.auth.AuthMethodPickerLayout
 import com.firebase.ui.auth.AuthUI
-import com.firebase.ui.auth.IdpResponse
+import com.firebase.ui.auth.FirebaseAuthUIActivityResultContract
 import com.firebase.ui.auth.data.model.FirebaseAuthUIAuthenticationResult
 import com.google.firebase.auth.FirebaseAuth
 
-class MainFragment : Fragment(){
+class MainFragment : Fragment() {
     companion object {
         const val TAG = "MainFragment"
         const val SIGN_IN_RESULT_CODE = 1001
@@ -28,6 +28,13 @@ class MainFragment : Fragment(){
     private val viewModel by viewModels<LoginViewModel>()
     private lateinit var _binding: FragmentMainBinding
     private val binding get() = _binding!!
+
+    // Firebase configuration
+    private val signInLauncher = registerForActivityResult(
+        FirebaseAuthUIActivityResultContract()
+    ) { res ->
+        this.onSignInResult(res)
+    }
 
     override fun onCreateView(
         inflater: LayoutInflater, container: ViewGroup?, savedInstanceState: Bundle?
@@ -48,10 +55,13 @@ class MainFragment : Fragment(){
                 // Launch SignInFLow
                 launchSignInFlow()
             } else {
+                val navController = findNavController()
+                navController.navigate(R.id.dashboardFragment)
                 // TODO: Navigate to the dashboard if logged in
             }
         })
     }
+
     private fun launchSignInFlow() {
         // Authentication providers
         val providers = arrayListOf(
@@ -80,7 +90,7 @@ class MainFragment : Fragment(){
         val response = result.idpResponse
         if (result.resultCode == Activity.RESULT_OK) {
             // successfully signed in
-            Log.d(StartupFragment.TAG,"Sign in successful")
+            Log.d(StartupFragment.TAG, "Sign in successful")
             var user = FirebaseAuth.getInstance().currentUser
         } else {
             val response = result.idpResponse
@@ -91,3 +101,12 @@ class MainFragment : Fragment(){
             }
         }
     }
+
+    private fun signOut() {
+        AuthUI.getInstance()
+            .signOut(requireContext())
+            .addOnCompleteListener{
+                Log.d(StartupFragment.TAG,"Signed out successfully")
+            }
+    }
+}
