+++ DungeonsandDragons/app/src/main/res/navigation/nav_graph.xml
@@ -3,29 +3,22 @@
     xmlns:app="http://schemas.android.com/apk/res-auto"
     xmlns:tools="http://schemas.android.com/tools"
     android:id="@+id/nav_graph.xml"
-    app:startDestination="@id/navigation">
-
-    <activity
-        android:id="@+id/startup"
-        android:name="com.example.dungeonsanddragons.StartupFragment"
-        android:label="Startup"
-        tools:layout="@layout/fragment_startup" />
-    <navigation android:id="@+id/navigation"
-        app:startDestination="@id/dashboard">
-        <fragment
-            android:id="@+id/dashboard"
-            android:name="com.example.dungeonsanddragons.DashboardFragment"
-            android:label="Dashboard"
-            tools:layout="@layout/fragment_dashboard">
-
-            <action
-                android:id="@+id/action_dashboard_to_NFC"
-                app:destination="@id/NFC" />
-        </fragment>
-        <fragment
-            android:id="@+id/NFC"
-            android:name="com.example.dungeonsanddragons.NFC"
-            android:label="NFC"
-            tools:layout="@layout/nfc" />
-    </navigation>
+    app:startDestination="@id/mainFragment">
+    <fragment
+        android:id="@+id/mainFragment"
+        android:name="com.example.dungeonsanddragons.MainFragment"
+        android:label="fragment_main"
+        tools:layout="@layout/fragment_main" >
+        <action
+            android:id="@+id/action_mainFragment_to_dashboardFragment"
+            app:destination="@id/dashboardFragment" />
+    </fragment>
+    <fragment
+        android:id="@+id/dashboardFragment"
+        android:name="com.example.dungeonsanddragons.DashboardFragment"
+        android:label="DashboardFragment" />
+    <fragment
+        android:id="@+id/nfcFragment"
+        android:name="com.example.dungeonsanddragons.NfcFragment"
+        android:label="NfcFragment" />
 </navigation>
